Rename repo states domain and dto mappers
