+++ extensions/ql-vscode/src/variant-analysis/repo-states-store/repo-states-domain-mapper.ts
